+++ packages/api-server/src/utils/misc.ts
@@ -1,7 +1,7 @@
 import { HOUR } from '@hugsmidjan/qj/time';
 import { format as _formatDate } from 'date-fns';
 import { is as locale } from 'date-fns/locale';
-import { FastifyReply } from 'fastify';
+import { FastifyReply, FastifyRequest } from 'fastify';
 import fs from 'fs';
 import { parse } from 'path';
 
@@ -19,7 +19,7 @@ export type Pms<keys extends string> = {
   Params: { [x in keys]: string };
 };
 /** Generates Querystring type declaration for Fastify's .get() method */
-export type QStr<keys extends string> = {
+export type QStr<keys extends string = string> = {
   Querystring: { [x in keys]?: string };
 };
 
@@ -31,6 +31,70 @@ export const cacheControl = (res: FastifyReply<any>, ttl_hrs: number): void => {
     'Cache-Control':
       'public, max-age=' + ttl_hrs * HOURS + (ttl_hrs ? ', immutable' : ''),
   });
+};
+
+// ---------------------------------------------------------------------------
+
+export const ensureObject = (cand: unknown): Record<string, unknown> => {
+  if (cand && typeof cand === 'object' && !Array.isArray(cand)) {
+    return cand as Record<string, unknown>;
+  }
+  return {};
+};
+export const ensureStringArray = (cand: unknown): Array<string> => {
+  if (Array.isArray(cand)) {
+    return cand.filter(
+      (item): item is string => !!item && typeof item === 'string',
+    );
+  }
+  return [];
+};
+
+// ---------------------------------------------------------------------------
+
+declare const _FileScopeToken__Brand: unique symbol;
+/** String that's safe to use as a folder/file name */
+export type FileScopeToken = string & { [_FileScopeToken__Brand]: true };
+
+export const ensureFileScopeToken = (
+  cand: unknown,
+): FileScopeToken | undefined => {
+  if (
+    typeof cand === 'string' &&
+    cand.length > 1 &&
+    /^[a-z0-9-](?:\.*[a-z0-9-]+)*$/i.test(cand)
+  ) {
+    return cand as FileScopeToken;
+  }
+};
+
+// ---------------------------------------------------------------------------
+
+const EMPTY_KEY = '_';
+const {
+  FILE_UPLOAD_KEY_DRAFT = EMPTY_KEY,
+  FILE_UPLOAD_KEY_PUBLISH = EMPTY_KEY,
+} = process.env;
+
+export type UploadType = 'draft' | 'publish';
+
+const apiKeyUsers: Record<string, UploadType | undefined> = {
+  [FILE_UPLOAD_KEY_DRAFT]: 'draft',
+  [FILE_UPLOAD_KEY_PUBLISH]: 'publish',
+};
+delete apiKeyUsers[EMPTY_KEY]; // Missing env keys must not open a security hole.
+
+/** Asserts that FastifyRequest.headers contain an allowed X-APIKey header value
+ * and returns the relevant UploadType string
+ *
+ * Throws if valid API key is not provided
+ */
+export const ensureUploadTypeHeader = (
+  req: Pick<FastifyRequest, 'headers'>,
+): UploadType | undefined => {
+  const apiKeyHeader = req.headers['X-APIKey'] || req.headers['x-apikey'];
+  const uploadType = apiKeyUsers[String(apiKeyHeader)];
+  return uploadType;
 };
 
 // ---------------------------------------------------------------------------